Move item preparation methods into a new ItemManager class
Reduces the magic located on the Component+Core protocol extension by
moving all item preparation and size calculation methods into a new
class called `ItemManager`. It resides on `ComponentManager` and is
used to set sizes on component model items.
